Update simple problem
added collision constraints by defining 3 sets of nodes/edges with corresponding aircraft that could possibly collide on that specific edge. (col/ot/ho)
this works for all nodes except for the start node which is constraint to be the same as the appearing time.
Next step is make the etv's dependent on the actual starting time, not the appearing time.